# apis/backend-api/skills/skills-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Skills Backend API",
    "version": "1.0.0",
    "x-sdkwork-owner": "sdkwork-skills",
    "x-sdkwork-api-authority": "sdkwork-skills-backend-api"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18092"
    }
  ],
...
  "paths": {
...
    "/backend/v3/api/skill_packages": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "skillPackages.list",
        "operationId": "skillPackages.list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillPackagesListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-skills",
        "x-sdkwork-api-authority": "sdkwork-skills-backend-api",
        "x-sdkwork-domain": "skills",
        "x-sdkwork-resource": "skillPackages",
        "x-sdkwork-public": false,
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-permission": "skills.packages.manage"
      },
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "skillPackages.create",
        "operationId": "skillPackages.create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillPackageCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillPackageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-skills",
        "x-sdkwork-api-authority": "sdkwork-skills-backend-api",
        "x-sdkwork-domain": "skills",
        "x-sdkwork-resource": "skillPackages",
        "x-sdkwork-public": false,
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-permission": "skills.packages.manage"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          },
          "i18nKey": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          }
        }
      },
...
      "SkillPackageRecord": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "uuid",
          "tenantId",
          "organizationId",
          "ownerUserId",
          "skillKey",
          "packageKey",
          "code",
          "displayName",
          "categories",
          "tags",
          "status",
          "visibility",
          "featured",
          "sortWeight",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64"
          },
          "uuid": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64"
          },
          "organizationId": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64"
          },
          "ownerUserId": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64"
          },
          "skillKey": {
            "type": "string",
            "pattern": "^skill\\.[a-z0-9_-]+(?:\\.[a-z0-9_-]+)*$"
          },
          "packageKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "disabled",
              "archived",
              "deleted"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "tenant",
              "organization",
              "public"
            ]
          },
          "featured": {
            "type": "boolean"
          },
          "sortWeight": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CreateSkillArtifactCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "versionLabel",
          "artifactRef",
          "checksumSha256",
          "invocationKind",
          "entrypoint"
        ],
        "properties": {
          "versionLabel": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "artifactRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048,
            "pattern": "^drive://spaces/[^/]+/nodes/[^/]+$"
          },
          "checksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "sizeBytes": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "minLength": 1,
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "u64",
            "nullable": true
          },
          "invocationKind": {
            "type": "string",
            "enum": [
              "local-workflow",
              "process-adapter",
              "mcp-tool",
              "kernel-provider"
            ]
          },
          "entrypoint": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "inputSchema": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "outputSchema": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "configSchema": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "defaultConfig": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "securityProfileId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "yanked"
            ],
            "default": "draft"
          },
          "capabilityKeys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
      "CreateSkillPackageCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "skillKey",
          "code",
          "displayName",
          "initialArtifact"
        ],
        "properties": {
          "skillKey": {
            "type": "string",
            "pattern": "^skill\\.[a-z0-9_-]+(?:\\.[a-z0-9_-]+)*$"
          },
          "packageKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "disabled",
              "archived",
              "deleted"
            ],
            "default": "draft"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "tenant",
              "organization",
              "public"
            ],
            "default": "tenant"
          },
          "featured": {
            "type": "boolean",
            "default": false
          },
          "sortWeight": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "initialArtifact": {
            "$ref": "#/components/schemas/CreateSkillArtifactCommand"
          }
        }
      },
...
      "SkillPackagesPageData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "pageInfo"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillPackageRecord"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "SkillPackagesListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SkillPackagesPageData"
              }
            }
          }
        ]
      },
      "SkillPackageResourceData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/SkillPackageRecord"
          }
        }
      },
      "SkillPackageResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SkillPackageResourceData"
              }
            }
          }
        ]
      },
...
    }
  },
...
}